Add support for getting video offset from motion photo v2 file formats.
Parse through all XMP nodes in the v2 XMP metadata and check if the nodes belong to the Directory array. If the property for the node matches the desired length or padding property paths, then add those values to the video offset.

Also, updated unit tests to reflect additional v2 format and refactored setUp() for modularity.