Refactor BufferProcessor into MotionPhotoReaderUtils.java
The reason for this refactor is that the BufferProcessor class shared many variables with MotionPhotoReader, so it was easier to write code if the process() function in BufferProcessor (which handled calls to nextFrame() and seekTo() synchronously) was moved back into the respective methods in MotionPhotoReader. The remaining methods that BufferProcessor contains also were not all necessarily related to processing the buffer queues, so it made more sense to make it a general reader utilities class.

- Move process() in BufferProcessor back into nextFrame() and seekTo().
- Move all shared variables back into MotionPhotoReader and pass them into the methods of MotionPhotoReaderUtils as arguments instead.
- Make all methods in MotionPhotoReaderUtils public static.
- Update MotionPhotoReaderTest to reflect the refactor.
- Create new tests under MotionPhotoUtilsTest (previously BufferProcessorTest).